Lint: fix all swiftlint violations in MistKit and MistDemo
Resolve file_name, line_length, file_length, multiline_arguments_brackets,
cyclomatic_complexity, and missing doc comment violations by splitting
large files and shortening deprecation messages.
