+++ Examples/MistDemo/Sources/MistDemoKit/Configuration/MistDemoConfig+DatabaseCredentials.swift
@@ -40,45 +40,60 @@ extension MistDemoConfig {
   internal func toDatabaseCredentials() throws -> DatabaseCredentials {
     switch database {
     case .public:
-      guard let keyID, !keyID.isEmpty else {
-        throw ConfigurationError.missingRequired(
-          "key.id",
-          suggestion: "Provide via CLOUDKIT_KEY_ID environment variable"
-        )
-      }
-      let material: PrivateKeyMaterial
-      if let raw = privateKey, !raw.isEmpty {
-        material = .raw(raw)
-      } else if let path = privateKeyFile, !path.isEmpty {
-        material = .file(path: path)
-      } else {
-        throw ConfigurationError.missingRequired(
-          "private.key",
-          suggestion: "Provide via CLOUDKIT_PRIVATE_KEY or CLOUDKIT_PRIVATE_KEY_PATH"
-        )
-      }
-      return .publicDatabase(keyID: keyID, privateKey: material)
-
+      return try toPublicCredentials()
     case .private, .shared:
-      let resolvedAPIToken = AuthenticationHelper.resolveAPIToken(apiToken)
-      guard !resolvedAPIToken.isEmpty else {
-        throw ConfigurationError.missingRequired(
-          "api.token",
-          suggestion: "Provide via CLOUDKIT_API_TOKEN environment variable"
-        )
-      }
-      let resolvedWebAuth = webAuthToken.flatMap {
-        AuthenticationHelper.resolveWebAuthToken($0)
-      }
-      guard let resolvedWebAuth, !resolvedWebAuth.isEmpty else {
-        throw ConfigurationError.missingRequired(
-          "web.auth.token",
-          suggestion: "Provide via CLOUDKIT_WEB_AUTH_TOKEN or run `mistdemo auth-token`"
-        )
-      }
-      return database == .private
-        ? .privateDatabase(apiToken: resolvedAPIToken, webAuthToken: resolvedWebAuth)
-        : .sharedDatabase(apiToken: resolvedAPIToken, webAuthToken: resolvedWebAuth)
+      return try toUserCredentials()
     }
+  }
+
+  private func toPublicCredentials() throws -> DatabaseCredentials {
+    guard let keyID, !keyID.isEmpty else {
+      throw ConfigurationError.missingRequired(
+        "key.id",
+        suggestion: "Provide via CLOUDKIT_KEY_ID environment variable"
+      )
+    }
+    let material = try resolvePrivateKeyMaterial()
+    return .publicDatabase(keyID: keyID, privateKey: material)
+  }
+
+  private func resolvePrivateKeyMaterial() throws -> PrivateKeyMaterial {
+    if let raw = privateKey, !raw.isEmpty {
+      return .raw(raw)
+    } else if let path = privateKeyFile, !path.isEmpty {
+      return .file(path: path)
+    }
+    throw ConfigurationError.missingRequired(
+      "private.key",
+      suggestion: "Provide via CLOUDKIT_PRIVATE_KEY or CLOUDKIT_PRIVATE_KEY_PATH"
+    )
+  }
+
+  private func toUserCredentials() throws -> DatabaseCredentials {
+    let resolvedAPIToken = AuthenticationHelper.resolveAPIToken(apiToken)
+    guard !resolvedAPIToken.isEmpty else {
+      throw ConfigurationError.missingRequired(
+        "api.token",
+        suggestion: "Provide via CLOUDKIT_API_TOKEN environment variable"
+      )
+    }
+    let resolvedWebAuth = webAuthToken.flatMap {
+      AuthenticationHelper.resolveWebAuthToken($0)
+    }
+    guard let resolvedWebAuth, !resolvedWebAuth.isEmpty else {
+      throw ConfigurationError.missingRequired(
+        "web.auth.token",
+        suggestion: "Provide via CLOUDKIT_WEB_AUTH_TOKEN or run `mistdemo auth-token`"
+      )
+    }
+    return database == .private
+      ? .privateDatabase(
+        apiToken: resolvedAPIToken,
+        webAuthToken: resolvedWebAuth
+      )
+      : .sharedDatabase(
+        apiToken: resolvedAPIToken,
+        webAuthToken: resolvedWebAuth
+      )
   }
 }
