+++ Sources/MistKit/Service/CloudKitService+Operations.swift
@@ -88,8 +88,7 @@ extension CloudKitService {
   ///   with `continuationMarker` or `queryAllRecords`
   @available(
     *, deprecated,
-    message:
-      "Use queryRecords(...) -> QueryResult to handle pagination explicitly, or queryAllRecords(...) to auto-paginate."
+    message: "Use queryRecords -> QueryResult for pagination, or queryAllRecords to auto-paginate."
   )
   public func queryRecords(
     recordType: String,
@@ -200,140 +199,6 @@ extension CloudKitService {
       return QueryResult(from: recordsData)
     } catch {
       throw mapToCloudKitError(error, context: "queryRecords")
-    }
-  }
-
-  /// Query all records, handling pagination automatically
-  ///
-  /// Convenience method that automatically fetches all matching records
-  /// by following continuation markers and making multiple requests if needed.
-  ///
-  /// - Parameters:
-  ///   - recordType: The type of records to query (must not be empty)
-  ///   - filters: Optional array of filters to apply to the query
-  ///   - sortBy: Optional array of sort descriptors
-  ///   - pageSize: Maximum number of records per page
-  ///     (1-200, defaults to `defaultQueryLimit`)
-  ///   - desiredKeys: Optional array of field names to fetch
-  ///   - maxPages: Maximum number of pages to fetch before throwing
-  ///     `CloudKitError.invalidResponse` (defaults to 1,000)
-  /// - Returns: Array of all matching records across all pages
-  /// - Throws: CloudKitError if any page request fails
-  ///
-  /// - Warning: Stops early if the server returns the same continuation
-  ///   marker with no new records (stuck-marker scenario), or if
-  ///   the page count exceeds `maxPages`.
-  public func queryAllRecords(
-    recordType: String,
-    filters: [QueryFilter]? = nil,
-    sortBy: [QuerySort]? = nil,
-    pageSize: Int? = nil,
-    desiredKeys: [String]? = nil,
-    maxPages: Int = 1_000
-  ) async throws(CloudKitError) -> [RecordInfo] {
-    var allRecords: [RecordInfo] = []
-    var currentMarker: String?
-    var pageCount = 0
-
-    repeat {
-      guard pageCount < maxPages else {
-        throw CloudKitError.paginationLimitExceeded(
-          maxPages: maxPages,
-          recordsCollected: allRecords.count
-        )
-      }
-
-      do {
-        try Task.checkCancellation()
-      } catch {
-        throw mapToCloudKitError(error, context: "queryAllRecords")
-      }
-
-      let result: QueryResult = try await queryRecords(
-        recordType: recordType,
-        filters: filters,
-        sortBy: sortBy,
-        limit: pageSize,
-        desiredKeys: desiredKeys,
-        continuationMarker: currentMarker
-      )
-
-      // Stuck-marker detection
-      if result.records.isEmpty && result.continuationMarker != nil
-        && result.continuationMarker == currentMarker
-      {
-        break
-      }
-
-      allRecords.append(contentsOf: result.records)
-      currentMarker = result.continuationMarker
-      pageCount += 1
-    } while currentMarker != nil
-
-    return allRecords
-  }
-
-  /// Modify (create, update, delete) records
-  @available(
-    *, deprecated,
-    message: "Use modifyRecords(_:) with RecordOperation in CloudKitService+WriteOperations instead"
-  )
-  internal func modifyRecords(
-    operations: [Components.Schemas.RecordOperation],
-    atomic: Bool = true
-  ) async throws(CloudKitError) -> [RecordInfo] {
-    do {
-      let response = try await client.modifyRecords(
-        .init(
-          path: createModifyRecordsPath(
-            containerIdentifier: containerIdentifier
-          ),
-          body: .json(
-            .init(
-              operations: operations,
-              atomic: atomic
-            )
-          )
-        )
-      )
-
-      let modifyData: Components.Schemas.ModifyResponse =
-        try await responseProcessor.processModifyRecordsResponse(response)
-      return modifyData.records?.compactMap { RecordInfo(from: $0) } ?? []
-    } catch {
-      throw mapToCloudKitError(error, context: "modifyRecords")
-    }
-  }
-
-  /// Lookup records by record names
-  public func lookupRecords(
-    recordNames: [String],
-    desiredKeys: [String]? = nil
-  ) async throws(CloudKitError) -> [RecordInfo] {
-    do {
-      let response = try await client.lookupRecords(
-        .init(
-          path: createLookupRecordsPath(
-            containerIdentifier: containerIdentifier
-          ),
-          body: .json(
-            .init(
-              records: recordNames.map { recordName in
-                .init(
-                  recordName: recordName,
-                  desiredKeys: desiredKeys
-                )
-              }
-            )
-          )
-        )
-      )
-
-      let lookupData: Components.Schemas.LookupResponse =
-        try await responseProcessor.processLookupRecordsResponse(response)
-      return lookupData.records?.compactMap { RecordInfo(from: $0) } ?? []
-    } catch {
-      throw mapToCloudKitError(error, context: "lookupRecords")
     }
   }
 }
