+++ Tests/MistKitTests/Service/CloudKitServiceFetchZoneChanges/CloudKitServiceTests.FetchZoneChanges+ErrorHandling.swift
@@ -98,7 +98,8 @@ extension CloudKitServiceTests.FetchZoneChanges {
         return
       }
       let service = try CloudKitServiceTests.makeService(
-        provider: ResponseProvider.connectionLost())
+        provider: ResponseProvider.connectionLost()
+      )
 
       await #expect {
         _ = try await service.fetchZoneChanges()
